Move missed file from Play namespace
