+++ osu.Game/Screens/Multi/Multiplayer/MultiplayerPlayer.cs
@@ -11,7 +11,7 @@ using osu.Game.Graphics.UserInterface;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Online.Rooms;
 using osu.Game.Scoring;
-using osu.Game.Screens.Multi.Play;
+using osu.Game.Screens.Multi.Playlists;
 using osu.Game.Screens.Play;
 using osu.Game.Screens.Play.HUD;
 using osu.Game.Screens.Ranking;
